+++ changelog.md
@@ -1,2 +1,5 @@
 # Changes
-- Updated cobblemon version to 1.5.
+- Generated Loot Balls are now a bit rarer.
+- Added 2 new configs:
+  - `minLootBallsPerChunk`: You can now set the minimum amount of lootballs generated per chunk. (Defaults to `0`)
+  - `generationChance`: You can now set the chance for a lootball generation attempt to be successful. (Defaults to `0.08`)